requirements.txt
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
six==1.16.0

pyserial~=3.5